refactor: 함수명 변경 및 파일 이동 (#16)
